+++ loan-service/src/main/java/com/example/loan_service/repository/DocumentRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.ArrayList;
 
 @Repository
-public interface DocumentRepository extends JpaRepository<DocumentEntity, Integer> {
+public interface DocumentRepository extends JpaRepository<DocumentEntity, Long> {
     ArrayList<DocumentEntity> findByLoanid(Long loanid);
 }
